+++ PSReminderLite.psm1
@@ -1,8 +1,8 @@
 
 #region Main
 
-Get-ChildItem -path $PSScriptRoot\functions\*.ps1 |
-ForEach-Object { . $_.FullName}
+Get-ChildItem -Path $PSScriptRoot\functions\*.ps1 |
+ForEach-Object { . $_.FullName }
 
 #endregion
 
@@ -13,17 +13,38 @@ If (Test-Path -Path $ExportPath) {
     #use the preference file
     Get-Content -Path $ExportPath | ConvertFrom-Json |
     ForEach-Object {
-        Set-Variable -name $_.Name -value $_.Value -Force
+        if ($_.Value.GetType().Name -ne "PSCustomObject") {
+            Set-Variable -Name $_.Name -Value $_.Value -Force
+        }
+        elseif ($_.name -eq 'PSReminderTag') {
+            $Hash = @{}
+            $global:t = $_
+            $_.Value.PSObject.Properties | ForEach-Object {
+                $Hash[$_.Name] = $_.Value
+            }
+            Set-Variable -Name PSReminderTag -Value $Hash -Force
+        }
     }
+<#
+Get-Content -Path $ExportPath | ConvertFrom-Json |
+    ForEach-Object {
+        Set-Variable -Name $_.Name -Value $_.Value -Force
+    }
+#>
 }
 else {
     #the default number of days to display for Show-TickleEvents
     $PSReminderDefaultDays = 7
 
     #database defaults
-    $PSReminderDB = Join-Path -path $HOME -ChildPath PSReminder.db
+    $PSReminderDB = Join-Path -Path $HOME -ChildPath PSReminder.db
     $PSReminderTable = 'EventData'
     $PSReminderArchiveTable = 'ArchivedEvent'
+    #define a default tag list and style settings
+    $PSReminderTag = @{
+        'Work'     = "`e[38;5;192m"
+        'Personal' = "`e[96m"
+    }
 }
 
 If (-Not (Test-Path -Path $PSReminderDB)) {
@@ -32,22 +53,24 @@ If (-Not (Test-Path -Path $PSReminderDB)) {
 
 #endregion
 
-#region Class definition
+#region Class definitions
 
 Class PSReminder {
     [String]$Event
     [DateTime]$Date
     [String]$Comment
     [int32]$ID
+    [string[]]$Tags
     [boolean]$Expired = $False
     hidden [TimeSpan]$Countdown
 
     #constructor
-    PSReminder([int32]$ID, [String]$Event, [DateTime]$Date, [String]$Comment) {
+    PSReminder([int32]$ID, [String]$Event, [DateTime]$Date, [String]$Comment, [String[]]$Tags) {
         $this.ID = $ID
         $this.Event = $Event
         $this.Date = $Date
         $this.Comment = $Comment
+        $this.Tags = $Tags
         if ($Date -lt (Get-Date)) {
             $this.Expired = $True
         }
@@ -64,29 +87,31 @@ Class ArchivePSReminder {
     [DateTime]$Date
     [String]$Comment
     [int32]$ID
+    [string[]]$Tags
     [DateTime]$ArchivedDate
 
-    ArchivePSReminder([int32]$ID, [String]$Event, [DateTime]$Date, [String]$Comment, [DateTime]$ArchivedDate) {
+    ArchivePSReminder([int32]$ID, [String]$Event, [DateTime]$Date, [String]$Comment, [DateTime]$ArchivedDate, [String[]]$Tags) {
         $this.ID = $ID
         $this.Event = $Event
         $this.Date = $Date
         $this.Comment = $Comment
         $this.ArchivedDate = $ArchivedDate
+        $this.Tags = $Tags
     }
 } #close ArchivePSReminder class
 
 Update-TypeData -TypeName PSReminder -DefaultDisplayPropertySet ID, Date, Event, Comment -Force
-Update-TypeData -TypeName PSReminder -MemberType AliasProperty -MemberName Name -Value Event -force
+Update-TypeData -TypeName PSReminder -MemberType AliasProperty -MemberName Name -Value Event -Force
 
 #endregion
 
 $export = @{
     Variable = @('PSReminderDefaultDays', 'PSReminderDB', 'PSReminderTable',
-    'PSReminderArchiveTable')
-    Function = @('Export-PSReminderPreference','Initialize-PSReminderDatabase',
-    'Add-PSReminder','Get-PSReminder','Get-PSReminderDBInformation','Set-PSReminder',
-    'Remove-PSReminder','Export-PSReminderDatabase','Import-PSReminderDatabase',
-    'Move-PSReminder','Get-AboutPSReminder')
-    Alias    = @('apsr','gpsr','spsr','rpsr','Archive-PSReminder')
+        'PSReminderArchiveTable','PSReminderTag')
+    Function = @('Export-PSReminderPreference', 'Initialize-PSReminderDatabase',
+        'Add-PSReminder', 'Get-PSReminder', 'Get-PSReminderDBInformation', 'Set-PSReminder',
+        'Remove-PSReminder', 'Export-PSReminderDatabase', 'Import-PSReminderDatabase',
+        'Move-PSReminder', 'Get-AboutPSReminder')
+    Alias    = @('apsr', 'gpsr', 'spsr', 'rpsr', 'Archive-PSReminder')
 }
 Export-ModuleMember @export